requirements.txt
flask==3.0.2
flask-cors==4.0.0
anthropic==0.18.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0
bcrypt==4.1.2
PyPDF2==3.0.1
gunicorn==21.2.0 